+++ packages/commerce-sdk-react/src/auth/index.ts
@@ -613,12 +613,14 @@ class Auth {
      * A wrapper method for commerce-sdk-isomorphic helper: logout.
      *
      */
-    async logout() {
+    async logout(options?: { isGuest?: boolean }) {
         // Not awaiting on purpose because there isn't much we can do if this fails.
-        void helpers.logout(this.client, {
-            accessToken: this.get('access_token'),
-            refreshToken: this.get('refresh_token_registered')
-        })
+        if(!options?.isGuest) {
+            void helpers.logout(this.client, {
+                accessToken: this.get('access_token'),
+                refreshToken: this.get('refresh_token_registered')
+            })
+        }
         this.clearStorage()
         return await this.loginGuestUser()
     }