+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v3.1.0-dev (Aug 08, 2024)
 
 - Add `defaultDnt` to support setting the dnt flag for SLAS. Upgrade `commerce-sdk-isomorphic` to v3.1.1 [#1979](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1979)
+- Add optional `isGuest` flag to logout helper []()
 
 ## v3.0.1 (Sep 04, 2024)
 - Fixed an issue where the `expires` attribute in cookies, ensuring it uses seconds instead of days. [#1994](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1994)