+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.1.0-exp-server-affinity.1 (Sep 11, 2024)
 ## v3.1.0-dev (Aug 08, 2024)
 
 - Add `defaultDnt` to support setting the dnt flag for SLAS. Upgrade `commerce-sdk-isomorphic` to v3.1.1 [#1979](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1979)